+++ src/request/trips-request/trips-request-params.ts
@@ -1,3 +1,5 @@
+import * as xmlbuilder from "xmlbuilder";
+
 import { TripLocationPoint } from "../../trip";
 import { IndividualTransportMode } from "../../types/individual-mode.types";
 import { TripModeType } from "../../types/trip-mode-type";
@@ -203,71 +205,8 @@ export class TripsRequestParams extends BaseRequestParams {
         }
       }
 
-      if (isMonomodal) {
-        if (isFrom) {
-          // https://github.com/openTdataCH/ojp-demo-app-src/issues/64
-          // Allow maxduration for more than 40m for walking / cycle monomodal routes
-          const modesWithOptions: IndividualTransportMode[] = ["walk", "cycle"];
-          if (modesWithOptions.indexOf(transportMode) !== -1) {
-            const transportModeOptionsNode = endPointNode.ele(
-              "IndividualTransportOptions"
-            );
-            transportModeOptionsNode.ele("Mode", transportMode);
-
-            if (transportMode === "walk") {
-              transportModeOptionsNode.ele("MaxDuration", "PT3000M");
-            }
-            if (transportMode === "cycle") {
-              transportModeOptionsNode.ele("MaxDuration", "PT600M");
-            }
-          }
-        }
-      } else {
-        const isOthersDriveCar =
-          transportMode === "taxi" || transportMode === "others-drive-car";
-        if (isOthersDriveCar) {
-          const hasExtension: boolean = (() => {
-            if (isFrom && this.modeType === "mode_at_end") {
-              return false;
-            }
-
-            if (!isFrom && this.modeType === "mode_at_start") {
-              return false;
-            }
-
-            return true;
-          })();
-
-          if (hasExtension) {
-            // TODO - in a method
-            const transportModeOptionsNode = endPointNode.ele(
-              "IndividualTransportOptions"
-            );
-            if (tripLocation.customTransportMode) {
-              transportModeOptionsNode.ele(
-                "Mode",
-                tripLocation.customTransportMode
-              );
-            }
-
-            transportModeOptionsNode.ele(
-              "MinDuration",
-              "PT" + tripLocation.minDuration + "M"
-            );
-            transportModeOptionsNode.ele(
-              "MaxDuration",
-              "PT" + tripLocation.maxDuration + "M"
-            );
-            transportModeOptionsNode.ele(
-              "MinDistance",
-              tripLocation.minDistance
-            );
-            transportModeOptionsNode.ele(
-              "MaxDistance",
-              tripLocation.maxDistance
-            );
-          }
-        }
+      if (isFrom) {
+        this.addAdditionalRestrictions(endPointNode, tripLocation);
       }
     });
 
@@ -312,6 +251,7 @@ export class TripsRequestParams extends BaseRequestParams {
 
     if (isMonomodal) {
       const standardModes: IndividualTransportMode[] = [
+        "foot",
         "walk",
         "self-drive-car",
         "cycle",
@@ -322,6 +262,7 @@ export class TripsRequestParams extends BaseRequestParams {
         paramsNode.ele("ItModesToCover", transportMode);
       }
 
+      // This is OJP v1
       const sharingModes: IndividualTransportMode[] = [
         "bicycle_rental",
         "car_sharing",
@@ -352,6 +293,10 @@ export class TripsRequestParams extends BaseRequestParams {
             ? this.fromTripLocation
             : this.toTripLocation;
 
+          if ((tripLocation.minDistance === null) || (tripLocation.maxDistance === null)) {
+            return;
+          }
+
           let tagName = isFrom ? "Origin" : "Destination";
           const endpointNode = paramsExtensionNode.ele(tagName);
 
@@ -374,5 +319,30 @@ export class TripsRequestParams extends BaseRequestParams {
     }
 
     paramsNode.ele("UseRealtimeData", 'explanatory');
+  }
+  
+  private addAdditionalRestrictions(nodeEl: xmlbuilder.XMLElement, tripLocation: TripLocationPoint) {
+    const hasAdditionalRestrictions = (tripLocation.minDuration !== null) || (tripLocation.maxDuration !== null) || (tripLocation.minDistance !== null) || (tripLocation.maxDistance !== null);
+    if (!hasAdditionalRestrictions) {
+      return;
+    }
+
+    const itNode = nodeEl.ele('IndividualTransportOptions');
+    if (tripLocation.customTransportMode) {
+      itNode.ele('Mode', tripLocation.customTransportMode);
+    }
+
+    if (tripLocation.minDuration !== null) {
+      itNode.ele('MinDuration', 'PT' + tripLocation.minDuration + 'M');
+    }
+    if (tripLocation.maxDuration !== null) {
+      itNode.ele('MaxDuration', 'PT' + tripLocation.maxDuration + 'M');
+    }
+    if (tripLocation.minDistance !== null) {
+      itNode.ele('MinDistance', tripLocation.minDistance);
+    }
+    if (tripLocation.maxDistance !== null) {
+      itNode.ele('MaxDistance', tripLocation.maxDistance);
+    }
   }
 }